+++ vert_groups.cpp
@@ -17,7 +17,7 @@ void VertGroups::add_one_group(const QList<unsigned int>& indices)
 	_groups.push_back(new VertGroup(indices));
 }
 
-int VertGroups::group_num() const
+size_t VertGroups::group_num() const
 {
 	return _groups.size();
 }
